+++ lora-phy/src/sx127x/radio_kind_params.rs
@@ -1,9 +1,6 @@
 use crate::mod_params::*;
 use crate::mod_traits::InterfaceVariant;
-use crate::sx127x::{
-    Sx127x, FREQUENCY_SYNTHESIZER_STEP, SX1272_RSSI_OFFSET, SX1276_RF_MID_BAND_THRESH, SX1276_RSSI_OFFSET_HF,
-    SX1276_RSSI_OFFSET_LF,
-};
+use crate::sx127x::Sx127x;
 use embedded_hal_async::spi::SpiDevice;
 
 /// Internal sx127x LoRa modes (signified by most significant bit flag)
@@ -324,281 +321,6 @@ pub trait Sx127xVariant {
     ) -> Result<(), RadioError>
     where
         Self: Sized;
-}
-
-/// Sx1272 implements the Sx127xVariant trait
-pub struct Sx1272;
-impl Sx127xVariant for Sx1272 {
-    fn bandwidth_value(bw: Bandwidth) -> Result<u8, RadioError> {
-        match bw {
-            Bandwidth::_125KHz => Ok(0x00),
-            Bandwidth::_250KHz => Ok(0x01),
-            Bandwidth::_500KHz => Ok(0x02),
-            _ => Err(RadioError::UnavailableBandwidth),
-        }
-    }
-
-    fn reg_txco() -> Register {
-        Register::RegTcxoSX1272
-    }
-
-    async fn set_tx_power<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        radio: &mut Sx127x<SPI, IV, Self>,
-        p_out: i32,
-        tx_boost: bool,
-    ) -> Result<(), RadioError> {
-        // SX1272 has two output pins:
-        // 1) RFO: (-1 to +14 dBm)
-        // 2) PA_BOOST: (+2 to +17 dBm and +5 to 20 +dBm)
-
-        // RegPaConfig - 0x32
-        // [7] - PaSelect (0: RFO, 1: PA_BOOST)
-        // [6:4] - Unused: 0
-        // [3:0] - Output power in dB steps
-
-        // RegPaDac - 0x5a (SX1272)
-        // [7:3] - Reserved (0x10 as default)
-        // [2:0] - PaDac: 0x04 default, 0x07 - enable +20 dBm on PA_BOOST
-
-        // TODO: Shall we also touch OCP settings?
-        if tx_boost {
-            // Deal with two ranges, +17dBm enables extra boost
-            if p_out > 17 {
-                // PA_BOOST out: +5 .. +20 dBm
-                let val = (p_out.min(20).max(5) - 5) as u8 & 0x0f;
-                radio.write_register(Register::RegPaConfig, (1 << 7) | val).await?;
-                radio.write_register(Register::RegPaDacSX1272, 0x87).await?;
-            } else {
-                // PA_BOOST out: +2 .. +17 dBm
-                let val = (p_out.min(17).max(2) - 2) as u8 & 0x0f;
-                radio.write_register(Register::RegPaConfig, (1 << 7) | val).await?;
-                radio.write_register(Register::RegPaDacSX1272, 0x84).await?;
-            }
-        } else {
-            // RFO out: -1 to +14 dBm
-            let val = (p_out.min(14).max(-1) + 1) as u8 & 0x0f;
-            radio.write_register(Register::RegPaConfig, val).await?;
-            radio.write_register(Register::RegPaDacSX1272, 0x84).await?;
-        }
-
-        Ok(())
-    }
-
-    fn ramp_value(ramp_time: RampTime) -> u8 {
-        // Sx1272 - default: 0x19
-        // [4]: LowPnTxPllOff - use higher power, lower phase noise PLL
-        //      only when the transmitter is used (default: 1)
-        //      0 - Standard PLL used in Rx mode, Lower PN PLL in Tx
-        //      1 - Standard PLL used in both Tx and Rx modes
-        ramp_time.value() | (1 << 4)
-    }
-
-    async fn set_modulation_params<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        radio: &mut Sx127x<SPI, IV, Self>,
-        mdltn_params: &ModulationParams,
-    ) -> Result<(), RadioError> {
-        let bw_val = Self::bandwidth_value(mdltn_params.bandwidth)?;
-        let sf_val = spreading_factor_value(mdltn_params.spreading_factor)?;
-
-        let cfg1 = radio.read_register(Register::RegModemConfig1).await?;
-        let ldro = mdltn_params.low_data_rate_optimize;
-        let cr_val = coding_rate_value(mdltn_params.coding_rate)?;
-        let val = (cfg1 & 0b110) | (bw_val << 6) | (cr_val << 3) | ldro;
-        radio.write_register(Register::RegModemConfig1, val).await?;
-        let cfg2 = radio.read_register(Register::RegModemConfig2).await?;
-        let val = (cfg2 & 0b1111) | (sf_val << 4);
-        radio.write_register(Register::RegModemConfig2, val).await?;
-        Ok(())
-    }
-
-    async fn set_packet_params<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        radio: &mut Sx127x<SPI, IV, Self>,
-        pkt_params: &PacketParams,
-    ) -> Result<(), RadioError>
-    where
-        Self: Sized,
-    {
-        let modemcfg1 = radio.read_register(Register::RegModemConfig1).await?;
-
-        let hdr = pkt_params.implicit_header as u8;
-        let crc = pkt_params.crc_on as u8;
-
-        let cfg1 = (modemcfg1 & 0b1111_1001) | (hdr << 2) | (crc << 1);
-        radio.write_register(Register::RegModemConfig1, cfg1).await?;
-        Ok(())
-    }
-
-    async fn rssi_offset<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        _: &mut Sx127x<SPI, IV, Self>,
-    ) -> Result<i16, RadioError> {
-        Ok(SX1272_RSSI_OFFSET)
-    }
-
-    async fn set_tx_continuous_wave_mode<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        _: &mut Sx127x<SPI, IV, Self>,
-    ) -> Result<(), RadioError> {
-        todo!()
-    }
-}
-
-/// Sx1276 implements the Sx127xVariant trait
-pub struct Sx1276;
-
-impl Sx127xVariant for Sx1276 {
-    fn bandwidth_value(bw: Bandwidth) -> Result<u8, RadioError> {
-        match bw {
-            Bandwidth::_7KHz => Ok(0x00),
-            Bandwidth::_10KHz => Ok(0x01),
-            Bandwidth::_15KHz => Ok(0x02),
-            Bandwidth::_20KHz => Ok(0x03),
-            Bandwidth::_31KHz => Ok(0x04),
-            Bandwidth::_41KHz => Ok(0x05),
-            Bandwidth::_62KHz => Ok(0x06),
-            Bandwidth::_125KHz => Ok(0x07),
-            Bandwidth::_250KHz => Ok(0x08),
-            Bandwidth::_500KHz => Ok(0x09),
-        }
-    }
-
-    fn reg_txco() -> Register {
-        Register::RegTcxoSX1276
-    }
-
-    async fn set_tx_power<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        radio: &mut Sx127x<SPI, IV, Self>,
-        p_out: i32,
-        tx_boost: bool,
-    ) -> Result<(), RadioError> {
-        let pa_reg = Register::RegPaDacSX1276;
-        if tx_boost {
-            // Output via PA_BOOST: [2, 20] dBm
-            let txp = p_out.min(20).max(2);
-
-            // Pout=17-(15-OutputPower)
-            let output_power: i32 = txp - 2;
-
-            if txp > 17 {
-                radio.write_register(pa_reg, PaDac::_20DbmOn.value()).await?;
-                radio.set_ocp(OcpTrim::_240Ma).await?;
-            } else {
-                radio.write_register(pa_reg, PaDac::_20DbmOff.value()).await?;
-                radio.set_ocp(OcpTrim::_100Ma).await?;
-            }
-            radio
-                .write_register(Register::RegPaConfig, PaConfig::PaBoost.value() | (output_power as u8))
-                .await?;
-        } else {
-            // Clamp output: [-4, 14] dBm
-            let txp = p_out.min(14).max(-4);
-
-            // Pmax=10.8+0.6*MaxPower, where MaxPower is set below as 7 and therefore Pmax is 15
-            // Pout=Pmax-(15-OutputPower)
-            let output_power: i32 = txp;
-
-            radio.write_register(pa_reg, PaDac::_20DbmOff.value()).await?;
-            radio.set_ocp(OcpTrim::_100Ma).await?;
-            radio
-                .write_register(
-                    Register::RegPaConfig,
-                    PaConfig::MaxPower7NoPaBoost.value() | (output_power as u8),
-                )
-                .await?;
-        }
-        Ok(())
-    }
-
-    fn ramp_value(ramp_time: RampTime) -> u8 {
-        // Sx1276 - default: 0x09
-        // [4]: reserved (0x00)
-        ramp_time as u8
-    }
-
-    async fn set_modulation_params<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        radio: &mut Sx127x<SPI, IV, Self>,
-        mdltn_params: &ModulationParams,
-    ) -> Result<(), RadioError> {
-        let bw_val = Self::bandwidth_value(mdltn_params.bandwidth)?;
-        let sf_val = spreading_factor_value(mdltn_params.spreading_factor)?;
-        let coding_rate_denominator_val = coding_rate_denominator_value(mdltn_params.coding_rate)?;
-
-        let mut config_2 = radio.read_register(Register::RegModemConfig2).await?;
-        config_2 = (config_2 & 0x0fu8) | ((sf_val << 4) & 0xf0u8);
-        radio.write_register(Register::RegModemConfig2, config_2).await?;
-
-        let mut config_1 = radio.read_register(Register::RegModemConfig1).await?;
-        config_1 = (config_1 & 0x0fu8) | (bw_val << 4);
-        radio.write_register(Register::RegModemConfig1, config_1).await?;
-
-        let cr = coding_rate_denominator_val - 4;
-        config_1 = radio.read_register(Register::RegModemConfig1).await?;
-        config_1 = (config_1 & 0xf1u8) | (cr << 1);
-        radio.write_register(Register::RegModemConfig1, config_1).await?;
-
-        let mut ldro_agc_auto_flags = 0x00u8; // LDRO and AGC Auto both off
-        if mdltn_params.low_data_rate_optimize != 0 {
-            ldro_agc_auto_flags = 0x08u8; // LDRO on and AGC Auto off
-        }
-        let mut config_3 = radio.read_register(Register::RegModemConfig3).await?;
-        config_3 = (config_3 & 0xf3u8) | ldro_agc_auto_flags;
-        radio.write_register(Register::RegModemConfig3, config_3).await
-    }
-
-    async fn set_packet_params<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        radio: &mut Sx127x<SPI, IV, Self>,
-        pkt_params: &PacketParams,
-    ) -> Result<(), RadioError> {
-        let mut config_1 = radio.read_register(Register::RegModemConfig1).await?;
-
-        if pkt_params.implicit_header {
-            config_1 |= 0x01u8;
-        } else {
-            config_1 &= 0xfeu8;
-        }
-        radio.write_register(Register::RegModemConfig1, config_1).await?;
-
-        let mut config_2 = radio.read_register(Register::RegModemConfig2).await?;
-        if pkt_params.crc_on {
-            config_2 |= 0x04u8;
-        } else {
-            config_2 &= 0xfbu8;
-        }
-        radio.write_register(Register::RegModemConfig2, config_2).await?;
-        Ok(())
-    }
-
-    async fn rssi_offset<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        radio: &mut Sx127x<SPI, IV, Self>,
-    ) -> Result<i16, RadioError> {
-        let frequency_in_hz = {
-            let msb = radio.read_register(Register::RegFrfMsb).await? as u32;
-            let mid = radio.read_register(Register::RegFrfMid).await? as u32;
-            let lsb = radio.read_register(Register::RegFrfLsb).await? as u32;
-            let frf = (msb << 16) + (mid << 8) + lsb;
-            (frf as f64 * FREQUENCY_SYNTHESIZER_STEP) as u32
-        };
-
-        if frequency_in_hz > SX1276_RF_MID_BAND_THRESH {
-            Ok(SX1276_RSSI_OFFSET_HF)
-        } else {
-            Ok(SX1276_RSSI_OFFSET_LF)
-        }
-    }
-
-    async fn set_tx_continuous_wave_mode<SPI: SpiDevice<u8>, IV: InterfaceVariant>(
-        radio: &mut Sx127x<SPI, IV, Self>,
-    ) -> Result<(), RadioError> {
-        radio.intf.iv.enable_rf_switch_rx().await?;
-        let pa_config = radio.read_register(Register::RegPaConfig).await?;
-        let new_pa_config = pa_config | 0b1000_0000;
-        radio.write_register(Register::RegPaConfig, new_pa_config).await?;
-        radio.write_register(Register::RegOpMode, 0b1100_0011).await?;
-        let modem_config = radio.read_register(Register::RegModemConfig2).await?;
-        let new_modem_config = modem_config | 0b0000_1000;
-        radio
-            .write_register(Register::RegModemConfig2, new_modem_config)
-            .await?;
-        Ok(())
-    }
 }
 
 #[allow(dead_code)]
